SAP_EWA_Analyzer_CrewAI/requirements.txt
# requirements.txt - CrewAI SAP EWA Analyzer (Fixed)
crewai>=0.28.0
crewai-tools>=0.8.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Web Interface
streamlit>=1.28.0

# PDF Processing
pdfplumber>=0.7.0
PyPDF2>=3.0.0
pypdf>=4.0.0

# Vector Store & AI
chromadb>=0.4.0
sentence-transformers>=2.2.2
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional PDF processing
pymupdf>=1.23.0

# Development (Optional)
pytest>=7.4.0
black>=23.12.0